Remove instance state op processor interface.
